Add random text selector
There are now 10 text files, and each will be randomly selected and
concatenated into a string. This string will be printed to the screen
through a set-size buffer and will break at newlines and truncate the
text at the end of the buffer.

I added an `offset` variable in order to make the text printable at
different offsets in the string. It should make more sense as
development continues.