Add downward text scrolling
The typing test can now continue for as long as the sample text string
is, although any text printed to the screen after the first screen of
text will not yet be properly formatted.

All instances of `curses.getsyx()` were replaced with `stdscr.getyx()`.

A non-ASCII character in text/04.txt was replaced with an ASCII-only
equivalent to make it easier for the user to type without a wrong
character penalty.